Move onKeepAlive to client/keepalive
